Rebuild for new fontconfig
I tried a test build to see whether this would unlock the Windows
build (as per an old comment in the meta.yaml), but it didn't work
trivially, and I'm not interested in spending any time on that
particular effort right now. It did look like it might not be very
hard to fix up the problem, though.